remove cache and introspect modules, add doc comments to pub methods
