+++ Cargo.toml
@@ -19,7 +19,7 @@ wasm-bindgen = "0.2"
 web-sys = { version = "0.3", optional = true, features = ["HtmlImageElement"]}
 
 tokio = { version = "1", features = ["rt-multi-thread", "rt", "fs"], optional = true }
-axum = { version = "0.7", optional = true }
+axum = { version = "0.7", optional = true, features = ["macros"] }
 tower = { version = "0.4", optional = true }
 tower-http = { version = "0.5", features = ["fs"], optional = true }
 
@@ -43,3 +43,6 @@ ssr = [
     "dep:futures", "dep:glob", 
 ]
 hydrate = [ "dep:web-sys","leptos/hydrate", "leptos_router/hydrate" ]
+
+[dev-dependencies]
+leptos_axum = "0.6.5"
