+++ src/lib.rs
@@ -1,17 +1,11 @@
 #![forbid(unsafe_code)]
 
-mod cache;
 mod image;
-mod introspect;
 mod optimizer;
 mod provider;
 mod routes;
 
-#[cfg(feature = "ssr")]
-pub use cache::*;
 pub use image::*;
-#[cfg(feature = "ssr")]
-pub use introspect::*;
 #[cfg(feature = "ssr")]
 pub use optimizer::ImageOptimizer;
 pub use provider::*;
