+++ src/provider.rs
@@ -1,18 +1,23 @@
 use crate::optimizer::CachedImage;
 use leptos::*;
 
-/// Provides Image Cache Context to the given scope.
+/// Provides Image Cache Context so that Images can use their blur placeholders if they exist.
+///
 /// This should go in the base of your Leptos <App/>.
 ///
-///Example
+/// Example
 ///
 /// ```
 /// use leptos_image::*;
+/// use leptos::*;
 ///
 /// #[component]
-/// pub fn App(cx: Scope) -> impl IntoView {
+/// pub fn App() -> impl IntoView {
 ///     provide_image_context();
 ///
+///     view!{
+///       <div/>
+///     }
 /// }
 ///
 /// ```
@@ -35,36 +40,6 @@ pub(crate) fn use_image_cache_resource() -> ImageResource {
     use_context::<ImageResource>().expect("Missing Image Resource")
 }
 
-#[cfg(feature = "ssr")]
-pub(crate) fn use_optimizer() -> Result<crate::ImageOptimizer, ServerFnError> {
-    use_context::<crate::ImageOptimizer>()
-        .ok_or_else(|| ServerFnError::ServerError("Image Optimizer Missing.".into()))
-}
-
-#[cfg(feature = "ssr")]
-pub(crate) async fn add_image_cache<I>(optimizer: &crate::optimizer::ImageOptimizer, images: I)
-where
-    I: IntoIterator<Item = CachedImage>,
-{
-    let images = images
-        .into_iter()
-        .filter(|image| matches!(image.option, crate::optimizer::CachedImageOption::Blur(_)))
-        .filter(|image| optimizer.cache.get(&image).is_none());
-
-    for image in images {
-        let path = optimizer.get_file_path_from_root(&image);
-        match tokio::fs::read_to_string(path).await {
-            Ok(data) => {
-                optimizer.cache.insert(image, data);
-                tracing::info!("Added image to cache with size {}", optimizer.cache.len())
-            }
-            Err(e) => {
-                tracing::error!("Failed to read image: {:?} with error: {:?}", image, e);
-            }
-        }
-    }
-}
-
 #[server(GetImageCache)]
 pub(crate) async fn get_image_cache() -> Result<Vec<(CachedImage, String)>, ServerFnError> {
     let optimizer = use_optimizer()?;
@@ -74,4 +49,10 @@ pub(crate) async fn get_image_cache() -> Result<Vec<(CachedImage, String)>, Serv
         .iter()
         .map(|entry| (entry.key().clone(), entry.value().clone()))
         .collect())
+}
+
+#[cfg(feature = "ssr")]
+pub(crate) fn use_optimizer() -> Result<crate::ImageOptimizer, ServerFnError> {
+    use_context::<crate::ImageOptimizer>()
+        .ok_or_else(|| ServerFnError::ServerError("Image Optimizer Missing.".into()))
 }
